As much as it pains me, we'll separate the audio from the cutscenes anyways :anguished:
I can't get the audio to not be f*cked one way or another when played in the video itself.
If I use the tweaks I did to Hxvlc, the last cutscene audio always get distorted to oblivion even when the audio is normalized just like the other cutscenes.
If I don't use the tweaks I did to Hxvlc, then audio is too quiet unless you play on like 5 bars of volume in the game, which leads to getting jumpscared when the songs countdown starts.

So we will have to do with this method... I hate it, but I can't be bothered to fix on this anymore for now.